add foreign id_lokasi to lokasi

+++ database/migrations/2021_07_03_080616_create_penginapans_table.php
@@ -17,7 +17,7 @@ class CreatePenginapansTable extends Migration
             $table->id('id_penginapan');
             $table->string('jenis_penginapan');
             $table->string('nama_penginapan');
-            $table->integer('id_lokasi');
+            $table->unsignedBigInteger('id_lokasi');
             $table->longText('deskripsi_penginapan');
         });
     }
